Made PPAPI examples buildable via gyp.
Also added audio example and fixed a few other examples.

BUG=none
TEST=all examples compile, audio plays music.

Review URL: http://codereview.chromium.org/3131010

+++ examples/stub/stub.cc
@@ -13,7 +13,7 @@ class MyInstance : public pp::Instance {
   explicit MyInstance(PP_Instance instance) : pp::Instance(instance) {}
   virtual ~MyInstance() {}
 
-  virtual bool Init(size_t argc, const char* argn[], const char* argv[]) {
+  virtual bool Init(uint32_t argc, const char* argn[], const char* argv[]) {
     return true;
   }
 };
